WIP: Added a StatusButton component
