+++ src/client/src/ui/status-connection/StatusButtonContainer.tsx
@@ -1,0 +1,15 @@
+import _ from 'lodash'
+import { connect } from 'react-redux'
+
+import { GlobalState } from 'StoreTypes'
+import { selectServices } from './selectors'
+import { StatusButton } from './StatusButton'
+
+const mapStateToProps = (state: GlobalState) => {
+  return {
+    connectionStatus: state.connectionStatus,
+    services: selectServices(state),
+  }
+}
+
+export default connect(mapStateToProps)(StatusButton)
