+++ src/client/src/ui/status-connection/index.ts
@@ -1,0 +1,1 @@
+export { default } from './StatusButtonContainer'
